Rescale the frictional heating (flux) from the flow solution units (MJ m-2 yr-1) to the units of flux corresponding to the enthalpy solution (J m-2 yr-1).
A copy of the `get_FrictionLoads` was added to the `src` folder,
with the additional support of a 'Friction Load Scaling Factor'
to deal with different units b/w the flow solution and the conservation
of energy. Added to makefile as well.

Comments were added to .sif to explain the conversion.