Pin the `netcdf` and `netcdf-fortan` library versions in the moudule file.
I'm now encoutering "NetcdfOutputSolver: Input/output error" as
reasons for simulations failing. This is new to the runs on the Graham
cluster, where I was using version `4.9.0` of the `netcdf` library,
whereas I was using version `4.7.4` on Cedar. The version of
`netcdf-fortan` was the same on Graham and Cedar (4.6.0), but I've pinned it
just to be safe. If this doesn't fix the failures then there may be some
small operating system difference between the clusters which is
resonsible.