Revert to minimal python env on cedar and swtich to python 3.11.5
Problems within needing to use the pre-built wheels on cedar, have
meant we need to strip back and usea a minimal python env on cedar,
instead of the dev-environment file.

Switched versions of python so it would be compatible with
StdEnv/2020, which is needed by elmerfem module. Better to have a
slightly older python version than multiple python environments.